Updated Readme.md with an example of how to initialize with SSL cert path.
I used Rails-specific code, which does make the readme less framework-agnostic.  Delete `Rails.application.config.middleware.` if this is undesired.

SSL instructions along these lines should be included with the readme as I just installed Omniauth on an app and got an SSL error, as described [on this StackOverflow thread](http://stackoverflow.com/questions/3977303/omniauth-facebook-certificate-verify-failed).  This is a solved issue and just needs to be documented.