Change debugging logging to use debug channel
These messages are very specific to omniauth internals, and thus are
only useful when debugging or developing. Having these messages on
`:info` causes them to fill test suite output or other non-production
logs which are at the `:info` level. We should move these messages to
`:debug` so developers can still see them, but they to not pollute
non-development logs.