Please make the implementation work in development
The suggested implementation doesn't work in development. It produces a few errors that are baffling to a newbie.
I'm here with a rails expert (@avdi) and it took the two of us a long time to get this working as far as you intended it to.

The development form issues a POST, not a GET. I don't yet know what the other sites do, but POST is necessary for omniauth to work in development.
And the verify_authentication_token check fails, so somehow this has to be worked around in order to trigger the create action.

Please please, for newbies, I don't know that these are the right changes, but make it so that people new to Rails and Omniauth can run in development mode immediately.